DELETE /posts/:id?version=n deletes post when specified version is the only version

+++ lib/tentd/api/posts.rb
@@ -304,12 +304,12 @@ module TentD
       class Destroy < Middleware
         def action(env)
           authorize_env!(env, :write_posts)
-          if env.params.has_key?(:version)
+          if env.params.has_key?(:version) && TentD::Model::PostVersion.qualify.join(:posts, :posts__id => :post_versions__post_id).count > 1
             if (post_version = TentD::Model::PostVersion.where(:user_id => Model::User.current.id, :post_id => env.params.post_id, :version => env.params.version.to_i).qualify.join(:posts, :posts__id => :post_versions__post_id).where(:posts__original => true).first) && post_version.destroy
               env.response = ''
             end
           else
-            if (post = TentD::Model::Post.first(:user_id => Model::User.current.id, :id => env.params.post_id, :original => true)) && post.destroy
+            if (post = TentD::Model::Post.first(:user_id => Model::User.current.id, :id => env.params.post_id, :original => true)) && (!env.params.has_key?(:version) || post.latest_version(:fields => [:version]).version == env.params.version.to_i) && post.destroy
               env.response = ''
               env.notify_deleted_post = post
             end
